+++ pytorch_lightning/core/decorators.py
@@ -10,7 +10,7 @@ def data_loader(fn):
     """
     wraps(fn)
     attr_name = '_lazy_' + fn.__name__
-
+    @wraps(fn)
     def _get_data_loader(self):
         try:
             value = getattr(self, attr_name)