+++ pytorch_lightning/core/lightning.py
@@ -869,7 +869,6 @@ class LightningModule(ABC, GradInformation, ModelIO, ModelHooks):
         return splits
 
     @data_loader
-    @abstractmethod
     def train_dataloader(self):
         """Implement a PyTorch DataLoader
 
@@ -895,8 +894,8 @@ class LightningModule(ABC, GradInformation, ModelIO, ModelHooks):
                 )
                 return loader
 
-
         """
+        return None
 
     @data_loader
     def tng_dataloader(self):  # todo: remove in v0.8.0